"Upload" page for DermAI
I have currently designed the basic structure of the application as that in the ppt, all the color codes, font styles etc.
Also added the "Pick Image from Gallery" and it works. Choosing a file will display it on the screen and if there is any buffer, a "loading..." text will be displaced.

The "Description" element doesn't really work properly as of now since it is supposed to start analyzing as soon as image is uploaded.

TODO:
- Add a navigation bar as similar as possible as that in the ppt.